requirements.txt
flask==2.0.1
werkzeug==2.0.1
pathlib==1.0.1
python-dotenv==0.19.1
PyJWT==2.3.0
Flask-SQLAlchemy==2.5.1
sqlalchemy==1.4.23
Flask-Migrate==3.1.0
flask-bcrypt==0.7.1
mysql-connector-python==8.0.26
flask-swagger-ui==4.11.1
gunicorn==20.1.0
flask-limiter==3.5.0
pytest==7.4.3
pytest-flask==1.3.0
pydantic>=2.5.0
